+++ changelog.md
@@ -5,6 +5,11 @@ This project follows semVer, where:
 - **[patch]** is a bugfix
 
 <!-- [planned] -->
+<!-- [unreleased] 
+- **[fix]** - inc/dec year issue on exact nye millisecond
+- **[change]** support parsing quarter-names as input - 'q2 2001'
+-->
+
 
 ### v6.8.0
 - **[fix]** major DST issue #182 (thanks Boris!)